# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/requisition_skills:
    get:
      tags:
        - RequisitionSkills
      summary: List RequisitionSkills
      description: Get a paginated list of RequisitionSkills with optional filtering and sorting
      operationId: listRequisitionSkills
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequisitionSkillList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RequisitionSkills
      summary: Create RequisitionSkill
      description: Create a new RequisitionSkill
      operationId: createRequisitionSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequisitionSkillRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequisitionSkill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    RequisitionSkill:
      type: object
      required:
        - id
        - requisition_id
        - skill_id
        - required_proficiency
        - metadata
      properties:
        id:
          type: string
          format: uuid
        requisition_id:
          type: string
          format: uuid
        skill_id:
          type: string
          format: uuid
        required_proficiency:
          type: string
        metadata:
          type: object
...
    CreateRequisitionSkillRequest:
      type: object
      required:
        - requisition_id
        - skill_id
        - required_proficiency
        - metadata
      properties:
        requisition_id:
          type: string
          format: uuid
        skill_id:
          type: string
          format: uuid
        required_proficiency:
          type: string
        metadata:
          type: object
...
    RequisitionSkillList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RequisitionSkill'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []