# src/domain/entity/requisition_skill.rs
use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use sqlx::FromRow;
use uuid::Uuid;

use super::ProficiencyLevel;
use super::AuditMetadata;

/// Strongly-typed ID for RequisitionSkill
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
#[serde(transparent)]
pub struct RequisitionSkillId(pub Uuid);

impl RequisitionSkillId {
    pub fn new(id: Uuid) -> Self { Self(id) }
    pub fn generate() -> Self { Self(Uuid::new_v4()) }
    pub fn into_inner(self) -> Uuid { self.0 }
}

impl std::fmt::Display for RequisitionSkillId {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        write!(f, "{}", self.0)
    }
}

impl std::str::FromStr for RequisitionSkillId {
    type Err = uuid::Error;
    fn from_str(s: &str) -> Result<Self, Self::Err> {
        Ok(Self(Uuid::parse_str(s)?))
    }
}

impl From<Uuid> for RequisitionSkillId {
    fn from(id: Uuid) -> Self { Self(id) }
}

impl From<RequisitionSkillId> for Uuid {
    fn from(id: RequisitionSkillId) -> Self { id.0 }
}

impl AsRef<Uuid> for RequisitionSkillId {
    fn as_ref(&self) -> &Uuid { &self.0 }
}

impl std::ops::Deref for RequisitionSkillId {
    type Target = Uuid;
    fn deref(&self) -> &Self::Target { &self.0 }
}

#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
pub struct RequisitionSkill {
    pub id: Uuid,
    pub requisition_id: Uuid,
    pub skill_id: Uuid,
    pub required_proficiency: ProficiencyLevel,
    #[serde(default)]
    #[sqlx(json)]
    pub metadata: AuditMetadata,
}

impl RequisitionSkill {
    /// Create a builder for RequisitionSkill
    pub fn builder() -> RequisitionSkillBuilder {
        <RequisitionSkillBuilder as Default>::default()
    }

    /// Create a new RequisitionSkill with required fields
    pub fn new(requisition_id: Uuid, skill_id: Uuid, required_proficiency: ProficiencyLevel) -> Self {
        Self {
            id: Uuid::new_v4(),
            requisition_id,
            skill_id,
            required_proficiency,
            metadata: AuditMetadata::default(),
        }
    }

    /// Get the entity's unique identifier
    pub fn id(&self) -> &Uuid {
        &self.id
    }

    /// Get a strongly-typed ID for this entity
    pub fn typed_id(&self) -> RequisitionSkillId {
        RequisitionSkillId(self.id)
    }

    /// Get when this entity was created
    pub fn created_at(&self) -> Option<&DateTime<Utc>> {
        self.metadata.created_at.as_ref()
    }

    /// Get when this entity was last updated
    pub fn updated_at(&self) -> Option<&DateTime<Utc>> {
        self.metadata.updated_at.as_ref()
    }

    /// Check if this entity is soft deleted
    pub fn is_deleted(&self) -> bool {
        self.metadata.deleted_at.is_some()
    }

    /// Check if this entity is active (not deleted)
    pub fn is_active(&self) -> bool {
        self.metadata.deleted_at.is_none()
    }

    /// Get when this entity was deleted
    pub fn deleted_at(&self) -> Option<&DateTime<Utc>> {
        self.metadata.deleted_at.as_ref()
    }

    /// Get who created this entity
    pub fn created_by(&self) -> Option<&Uuid> {
        self.metadata.created_by.as_ref()
    }

    /// Get who last updated this entity
    pub fn updated_by(&self) -> Option<&Uuid> {
        self.metadata.updated_by.as_ref()
    }

    /// Get who deleted this entity
    pub fn deleted_by(&self) -> Option<&Uuid> {
        self.metadata.deleted_by.as_ref()
    }


    // ==========================================================
    // Partial Update
    // ==========================================================

    /// Apply partial updates from a map of field name to JSON value
    pub fn apply_patch(&mut self, fields: std::collections::HashMap<String, serde_json::Value>) {
        for (key, value) in fields {
            match key.as_str() {
                "requisition_id" => {
                    if let Ok(v) = serde_json::from_value(value) { self.requisition_id = v; }
                }
                "skill_id" => {
                    if let Ok(v) = serde_json::from_value(value) { self.skill_id = v; }
                }
                "required_proficiency" => {
                    if let Ok(v) = serde_json::from_value(value) { self.required_proficiency = v; }
                }
                _ => {} // ignore unknown fields
            }
        }
    }

    // <<< CUSTOM METHODS START >>>
    // <<< CUSTOM METHODS END >>>
}

impl super::Entity for RequisitionSkill {
    type Id = Uuid;

    fn entity_id(&self) -> &Self::Id {
        &self.id
    }

    fn entity_type() -> &'static str {
        "RequisitionSkill"
    }
}

impl backbone_core::PersistentEntity for RequisitionSkill {
    fn entity_id(&self) -> String {
        self.id.to_string()
    }
    fn set_entity_id(&mut self, id: String) {
        if let Ok(uuid) = uuid::Uuid::parse_str(&id) {
            self.id = uuid;
        }
    }
    fn created_at(&self) -> Option<chrono::DateTime<chrono::Utc>> {
        self.metadata.created_at
    }
    fn set_created_at(&mut self, ts: chrono::DateTime<chrono::Utc>) {
        self.metadata.created_at = Some(ts);
    }
    fn updated_at(&self) -> Option<chrono::DateTime<chrono::Utc>> {
        self.metadata.updated_at
    }
    fn set_updated_at(&mut self, ts: chrono::DateTime<chrono::Utc>) {
        self.metadata.updated_at = Some(ts);
    }
    fn deleted_at(&self) -> Option<chrono::DateTime<chrono::Utc>> {
        self.metadata.deleted_at
    }
    fn set_deleted_at(&mut self, ts: Option<chrono::DateTime<chrono::Utc>>) {
        self.metadata.deleted_at = ts;
    }
}

impl backbone_orm::EntityRepoMeta for RequisitionSkill {
    fn column_types() -> std::collections::HashMap<String, String> {
        let mut m = std::collections::HashMap::new();
        m.insert("id".to_string(), "uuid".to_string());
        m.insert("requisition_id".to_string(), "uuid".to_string());
        m.insert("skill_id".to_string(), "uuid".to_string());
        m.insert("required_proficiency".to_string(), "proficiency_level".to_string());
        m
    }
    fn search_fields() -> &'static [&'static str] {
        &[]
    }
}

/// Builder for RequisitionSkill entity
///
/// Provides a fluent API for constructing RequisitionSkill instances.
/// System fields (id, metadata, timestamps) are auto-initialized.
#[derive(Debug, Clone, Default)]
pub struct RequisitionSkillBuilder {
    requisition_id: Option<Uuid>,
    skill_id: Option<Uuid>,
    required_proficiency: Option<ProficiencyLevel>,
}

impl RequisitionSkillBuilder {
    /// Set the requisition_id field (required)
    pub fn requisition_id(mut self, value: Uuid) -> Self {
        self.requisition_id = Some(value);
        self
    }

    /// Set the skill_id field (required)
    pub fn skill_id(mut self, value: Uuid) -> Self {
        self.skill_id = Some(value);
        self
    }

    /// Set the required_proficiency field (required)
    pub fn required_proficiency(mut self, value: ProficiencyLevel) -> Self {
        self.required_proficiency = Some(value);
        self
    }

    /// Build the RequisitionSkill entity
    ///
    /// Returns Err if any required field without a default is missing.
    pub fn build(self) -> Result<RequisitionSkill, String> {
        let requisition_id = self.requisition_id.ok_or_else(|| "requisition_id is required".to_string())?;
        let skill_id = self.skill_id.ok_or_else(|| "skill_id is required".to_string())?;
        let required_proficiency = self.required_proficiency.ok_or_else(|| "required_proficiency is required".to_string())?;

        Ok(RequisitionSkill {
            id: Uuid::new_v4(),
            requisition_id,
            skill_id,
            required_proficiency,
            metadata: AuditMetadata::default(),
        })
    }
}